Move ColorPickerIcon to separate file
